+++ samples/browser/scripts/module_with_arguments.js
@@ -1,0 +1,15 @@
+ (function () { 
+
+  "use strict";
+
+  console.log("=== Module with arguments sample: ===");
+
+  var util = $module(function (exports, global) {
+
+    global.myVariable = 42; 
+
+  }, [window]);
+
+  console.log(myVariable);
+
+}());