+++ src/module.js
@@ -9,14 +9,24 @@ var Module = Object.create(null, {
    * @static
    * @public
    * @method create
+   * @param {Function} arguments
    * @param {Function} scope
+   * @param {Array} args
    * @returns {Object}
    */
   create: {
-    value: function (scope) {
+    value: function (scope, args) {
       var module = {};
       if (scope) {
-        scope(module);
+        if (args) {
+          if (!(args instanceof Array)) {
+            throw new TypeException("Type of parameter 'args' is invalid");
+          }
+          args.splice(0, 0, module);
+          scope.apply(this, args); 
+        } else {
+          scope(module);
+        }
       }
       return module;
     },