+++ src/type_extension.js
@@ -1,0 +1,59 @@
+/**
+ * @class {class4js.TypeExtension}
+ * @constructor {class4js.TypeExtension}
+ * @param {Object} target
+ * @param {String} name
+ * @param {Function} value
+ */
+var TypeExtension = function (target, name, value) {
+  this.__target = target;
+  this.__name = name;
+  this.__value = value;
+};
+
+TypeExtension.prototype = Object.create(Object.prototype, {
+
+  /**
+   * @memberOf {class4js.TypeExtension}
+   * @public
+   * @property {Object} target
+   */
+  target: {
+    get: function () {
+      return this.__target;
+    },
+    enumerable: true,
+    configurable: false
+  },
+
+  /**
+   * @memberOf {class4js.TypeExtension}
+   * @public
+   * @property {String} name
+   */
+  name: {
+    get: function () {
+      return this.__name;
+    },
+    enumerable: true,
+    configurable: false
+  },
+
+  /**
+   * @memberOf {class4js.TypeExtension}
+   * @public
+   * @property {Function} value
+   */
+  value: {
+    get: function () {
+      return this.__value;
+    },
+    enumerable: true,
+    configurable: false
+  }
+
+});
+
+Object.seal(TypeExtension);
+Object.seal(TypeExtension.prototype);
+